requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
idna==3.4
numpy==1.23.5
Pillow==9.3.0
pip==22.3.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
setuptools==65.6.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
wheel==0.38.4
jsondiff~=2.0.0